+++ src/main/java/com/calculadora/complexa/Operacoes.java
@@ -2,6 +2,6 @@ package com.calculadora.complexa;
 
 public interface Operacoes {
 
-    float calcular ();
+    double calcular ();
 
 }
